MS_AccountManager/requirements.txt
cherrypy==18.8.0
requests==2.28.1
python-dotenv==0.21.0
pyjwt==2.6.0
bcrypt==4.0.1